pyproject.toml
[project]
name = "voicenode"
version = "0.1.0"
description = "Python client for distributed voice control"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4.6",
    "websockets>=12.0",
    "pywhispercpp>=1.2.0",
    "structlog>=24.1.0",
    "setuptools>=82.0.1",
    "webrtcvad-wheels>=2.0.14",
]

[project.scripts]
voicenode = "voicenode.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
